requirements.txt
# Core Computer Vision Libraries
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0

# Deep Learning and AI
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Image Processing
Pillow>=9.0.0
scikit-image>=0.20.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.0

# Data Science and Analysis
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.11.0
plotly>=5.10.0

# Tracking Libraries
filterpy>=1.4.0
lap>=0.4.0

# Video Processing
ffmpeg-python>=0.2.0

# Configuration and Utilities
PyYAML>=6.0
tqdm>=4.60.0
click>=8.0.0
loguru>=0.6.0

# Jupyter and Development
jupyter==1.0.0
notebook==7.0.2
ipywidgets==8.0.7

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Optional GPU acceleration (uncomment if using CUDA)
# torch==2.0.1+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision==0.15.2+cu118 --index-url https://download.pytorch.org/whl/cu118
